Linux fixes: Upgrade TargetFrameworkVerison to 4.6.2 and make it consistant across all projects.
MonoDevelop errors out when linking projects with different target versions. So we need the targets to be consistant across all projects.

Current MySql.Data nuget packages don't have a net40 build in them anymore. We need at least net462 for support, so net462 is what we need to go with.

Remove the startup tag from the various App.config files. (Once again broken target versions, and the system should be able to figure out what is needed here on it's own.)

Also, introduce the <TargetFramework> tag in the project files. (Needed for future target upgrades. As TargetFrameworkVerison is deprecated, but we don't remove it because MonoDevelop won't recognize the new TargetFramework tag. (It defaults to .NETFramework,v4.0 in that case. Which won't build due to missing nuget packages and target version conflicts.))